pyproject.toml
[tool.poetry]
name = "backup-site"
version = "0.1.0"
description = "Solution CLI de sauvegarde de site web avec support Docker"
authors = ["Julien <julien@example.com>"]
readme = "README.md"
packages = [{include = "backup_site", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
pyyaml = "^6.0.1"
cryptography = "^41.0.7"
paramiko = "^3.4.0"
rich = "^13.7.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[tool.poetry.scripts]
backup-site = "backup_site.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
